Restructure and have threads for each monitoring
